feat: Add command line argument support for calculating gcd of multiple numbers
This commit adds support for passing command line arguments to the program in order to calculate the greatest common divisor (gcd) of multiple numbers. The program now reads the arguments provided by the user and stores them in a vector. It then calculates the gcd of the numbers using the existing gcd function and prints the result.

Co-authored-by: [Author Name]